Refactored to separate concerns
Measurement became PhysicalQuantity, and the definition of units
was taken out of a config array in classes that implement Measurement
and moved to a UnitOfMeasure class and interface set.

This allows different Units of Measure to be implemented in different
ways, depending on how their conversion happens.  For example,
some units of measure conversions are simple factors, while others
involve more complicated behavior (like temperature).  Currency could
be a real time API call.

Right now, all Units of Measure are implemented the same way -
an internal set of closures.